requirements.txt
click~=6.7
netifaces
pygments
scapy-python3==0.23
termcolor
websockets~=4.0.0
